+++ Exporter/src/main/java/uk/ac/ebi/biostd/exporter/jobs/pmc/importer/process/PmcFileManager.java
@@ -7,20 +7,25 @@ import java.nio.file.Paths;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import lombok.AllArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
+import org.springframework.stereotype.Component;
+import uk.ac.ebi.biostd.exporter.jobs.pmc.importer.PmcImportProperties;
 import uk.ac.ebi.biostd.in.pagetab.SubmissionInfo;
 
 @Slf4j
-@AllArgsConstructor
+@Component
 public class PmcFileManager {
 
     private final static String DOWNLOAD_URL = "http://europepmc.org/articles/%s/bin/%s";
     private final ExecutorService executor = Executors.newFixedThreadPool(10);
 
     private final String userPath;
+
+    public PmcFileManager(PmcImportProperties properties) {
+        this.userPath = properties.getSubmitterUserPath();
+    }
 
     @SneakyThrows
     public void downloadFiles(SubmissionInfo submissionInfo) {
